trier-integration/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
SQLAlchemy==2.0.32
python-dotenv==1.0.1
